+++ src/components/AiSlider.tsx
